smp-affinity.pl: add possibility to configure split RX/TX IRQs
Most IRQs still have a single IRQ for both RX and TX, as well
as for Link Management.

However, an increasingly number of NICs has a dedicated IRQ
for both RX and for TX.  This e. g. is from e1000e:

 51:   11940158   11941022   PCI-MSI-edge      eth0-rx-0
 52:      94145      94258   PCI-MSI-edge      eth0-tx-0
 53:        292        131   PCI-MSI-edge      eth0

The last one is for Link Management.

Change the code such that

1) different RX and TX IRQs are properly supported
2) RX and TX IRQs can be configured differently if need be

The last one is achieved by adding two new values to the config:
RX_IRQ_CPUS and TX_IRQ_CPUS.

If TX_IRQ_CPUS is specified it takes precedence over any
value specified in IRQ_CPUS or CPUS.  And the same is true
for RX_IRQ_CPUS.

Signed-off-by: Holger Eitzenberger <holger.eitzenberger@sophos.com>