smp-affinity.pl: allow to specify different RX/TX mask
Some of the NICs have different IRQs for RX and TX.  This e. g.
is for some e1000e NICs:

 51:   12040572   14135921   PCI-MSI-edge      eth0-rx-0
 52:      95114     105376   PCI-MSI-edge      eth0-tx-0
 53:        292        131   PCI-MSI-edge      eth0

Such a NIC is now reported as:

  {
    'LM_IRQ' => '53',
    'NAME' => 'eth0',
    'QUEUES' => [
                  {
                    'NUM' => 0,
                    'RX_IRQ_MASK' => 2,
                    'TX_IRQ' => '52',
                    'TX_IRQ_MASK' => 3,
                    'RX_IRQ' => '51',
                    'STEERING_MASK' => 3
                  }
                ]
        };

The Link Management IRQ is reported as LM_IRQ once per NIC.  For each
queue the values RX_IRQ_MASK, TX_IRQ_MASK and STEERING_MASK
are reported.

Signed-off-by: Holger Eitzenberger <holger.eitzenberger@sophos.com>